fix double input

+++ models_torch.py
@@ -135,8 +135,8 @@ class CriticNetwork(nn.Module):
 class HRP(nn.Module):
     def __init__(self, target_model=False):
         super(HRP, self).__init__()
-        self.an = ActorNetwork()
-        self.cn = CriticNetwork()
+        self.an = ActorNetwork().double()
+        self.cn = CriticNetwork().double()
         self.critic_criterion = nn.MSELoss()
         self.discount_rate = 0.99
         self.learning_rate = 1e-4
